using token middleware for protection Meeting Controller

+++ app/Http/Controllers/MeetingController.php
@@ -6,13 +6,18 @@ use Illuminate\Http\Request;
 use App\Meeting;
 use App\Http\Requests;
 use Carbon\Carbon;
+use JWTAuth;
 
 class MeetingController extends Controller
 {
 
     public function __construct()
     {
-
+        $this->middleware('jwt.auth', ['only' => [
+            'update',
+            'store',
+            'destroy',
+        ]]);
     }
 
     /**
@@ -51,13 +56,16 @@ class MeetingController extends Controller
             'title'       => 'required',
             'description' => 'required',
             'time'        => 'required|date_format:YmdHie',
-            'userID'      => 'required'
         ]);
+
+        if (!$user = JWTAuth::parseToken()->authenticate()) {
+            return response()->json(['msg' => 'User not found'], 404);
+        }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $userId = $request->input('userID');
+        $userId = $user->id;
 
         $meeting = new Meeting([
             'time'        => Carbon::createFromFormat('YmdHie', $time),
@@ -111,36 +119,30 @@ class MeetingController extends Controller
      */
     public function update(Request $request, $id)
     {
+
         $this->validate($request, [
             'title'       => 'required',
             'description' => 'required',
             'time'        => 'required|date_format:YmdHie',
-            'userID'      => 'required'
         ]);
+
+        if (!$user = JWTAuth::parseToken()->authenticate()) {
+            return response()->json(['msg' => 'User not found'], 404);
+        }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $userId = $request->input('userID');
-        $meeting = [
-            'title'        => $title,
-            'description'  => $description,
-            'time'         => $time,
-            'user_id'      => $userId,
-            'view_meeting' => [
-                'href'   => 'api/v1/meeting/1',
-                'method' => 'GET'
-            ]
-        ];
+        $userId = $user->id;
 
         $meeting = Meeting::with('users')->findOrFail($id);
-
-        if(!$meeting->users()->where('users.id', $userId)->first())
+        if (!$meeting->users()->where('users.id', $userId)->first())
         {
             return response()->json(['msg' => 'user not registered for meeting, update not successful'],404);
         }
         $meeting->time = Carbon::createFromFormat('YmdHie', $time);
         $meeting->title = $title;
+        $meeting->description = $description;
         if (!$meeting->update()) {
             return response()->json(['msg' => 'Error during updating'], 404);
         }
@@ -167,6 +169,13 @@ class MeetingController extends Controller
     public function destroy($id)
     {
         $meeting = Meeting::findOrFail($id);
+        if (!$user = JWTAuth::parseToken()->authenticate()) {
+            return response()->json(['msg' => 'User not found'], 404);
+        }
+        if (!$meeting->users()->where('users.id', $user->id)->first())
+        {
+            return response()->json(['msg' => 'user not registered for meeting, update not successful'],404);
+        }
         $users = $meeting->users;
         $meeting->users()->detach();
         if (!$meeting->delete()) {
